+++ voting/frontend/app/layout.tsx
@@ -1,6 +1,5 @@
 import './globals.css';
 import '@rainbow-me/rainbowkit/styles.css';
-
 import { Providers } from './providers';
 import Headers from "@/components/headers/headers";