+++ voting/frontend/components/home/ShowResult.tsx
@@ -1,0 +1,11 @@
+import WinningProposal from "@/components/WinningProposal";
+
+export const ShowResult = () => {
+    return (
+        <>
+            <div className="pt-10">
+                <WinningProposal />
+            </div>
+        </>
+    );
+};