+++ voting/frontend/components/voters/SendVote.tsx
@@ -1,16 +1,19 @@
 "use client"
 
 import { writeContractByFunctionName } from "@/utils"
-import { useState } from "react"
-import { useToast } from "@chakra-ui/react"
-import Event from "../Event"
+import { useToast } from "@chakra-ui/react";
+import { votingSessionStartedStatus, WorkflowStatus } from "@/constants";
+import { useWorkflowStatusContext } from "@/context/workflowStatus";
+import GetProposalsList from "@/components/voters/GetProposalsList";
+import { useSelectProposalContext } from "@/context/SelectProposal";
 
 const SendVote = () => {
-    const [vote, setVote] = useState('')
+    const { workflowStatus } = useWorkflowStatusContext()
     const toast = useToast()
+    const {description, id} = useSelectProposalContext()
 
     const submitVote = () => {
-        if (vote.trim().length <= 0) {
+        if (id.trim().length <= 0) {
             toast({
                 title: 'Invalid Vote',
                 description: 'Vote can\'t be empty',
@@ -19,10 +22,10 @@ const SendVote = () => {
                 isClosable: true,
             })
         } else {
-            writeContractByFunctionName('setVote', vote).then(
+            writeContractByFunctionName('setVote', id).then(
                 () => toast({
-                    title: 'Voted',
-                    description: `You're vote is registered.`,
+                    title: 'Voted!',
+                    description: `Your vote has been sent. Thank you.`,
                     status: 'success',
                     duration: 5000,
                     isClosable: true,
@@ -39,31 +42,26 @@ const SendVote = () => {
         }
     }
 
-    return (<>
-        <section className="m-2 mx-auto w-3/4 rounded h-auto bg-gradient-to-r from-indigo-900 to-indigo-800 text-indigo-100 shadow-lg drop-shadow-lg border-indigo-600 border">
-            <h2 className="font-bold text-lg text-center mb-3">Vote for a Proposal</h2>
-            <form className="w-full m-auto p-3 pr-3">
-                <div className="flex justify-center items-center border-b border-indigo-300 py-2">
-                    <input className="mr-3 p-2 leading-tight focus:outline-none bg-indigo-100 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-indigo-700 focus:border-indigo-900"
-                        name="vote"
-                        required
-                        type="number"
-                        placeholder="Vote for a proposal..."
-                        value={vote}
-                        onChange={e => setVote(e.target.value)}
-                    />
+    return (
+        <>
+            <section className="m-2 mx-auto w-3/4 p-4 mb-4 text-sm rounded h-auto bg-gradient-to-r from-indigo-100 to-indigo-200 text-gray-900 shadow-lg drop-shadow-lg border-indigo-600 border">
 
-                    <button className="content-center bg-indigo-950 hover:bg-indigo-100 hover:text-gray-900 text-white font-semibold py-2 px-4 rounded-lg"
-                        type="button"
-                        onClick={() => submitVote()}
-                    >
-                        Send Vote
-                    </button>
-                </div>
-            </form>
-        </section>
-        
-        <Event name='Voted'></Event>
-    </>);
+                <GetProposalsList/>
+
+                {WorkflowStatus[workflowStatus] === votingSessionStartedStatus &&
+                    <form className="flex flex-col">
+                        <div className="p-2 w-3/4 mx-auto m-1 border-t border-indigo-300 py-2">
+                            <button className="w-full content-center bg-rose-400 hover:bg-rose-500 hover:text-gray-900 text-indigo-100 font-semibold text-lg py-6 px-4 rounded-lg"
+                                    type="button"
+                                    onClick={() => submitVote()}
+                            >Click to confirm your choice:  {description}
+                            </button>
+                        </div>
+                    </form>
+                }
+
+            </section>
+        </>
+    );
 }
 export default SendVote;