+++ voting/frontend/utils/index.ts
@@ -1,8 +1,9 @@
-import { abi, contractAddress, network } from "@/constants"
+import { abi, contractAddress, genesisBlock, network, ProposalsRegistered} from "@/constants"
 import { Voter } from "@/interfaces/Voter"
 import { readContract, prepareWriteContract, writeContract } from "@wagmi/core"
-import { BaseError, ContractFunctionRevertedError, createPublicClient, http } from "viem"
+import { BaseError, ContractFunctionRevertedError, createPublicClient, http, parseAbiItem } from "viem"
 import { hardhat, sepolia } from "viem/chains"
+import { Proposal } from "@/interfaces/Proposal";
 
 export const getWorkflowStatus = async (address: `0x${string}`): Promise<number> => {
     return readContractByFunctionName<number>('workflowStatus', address )
@@ -30,6 +31,12 @@ export const userIsVoter = async (address: `0x${string}`): Promise<boolean> => {
     return readContractByFunctionName<Voter>('getVoter', address, address).then(
         data => data && data.isRegistered
     ).catch(() => false)
+}
+
+export const GetProposals = async (address: `0x${string}`, logId: number): Promise<Proposal> => {
+    return readContractByFunctionName<Proposal>('getOneProposal', address, logId).then(
+        proposal => proposal
+    )
 }
 
 export const readContractByFunctionName = async <T>(functionName: string, address: `0x${string}`, ...args: `0x${string}`[]|string[]|number[]): Promise<T> => {
@@ -65,6 +72,20 @@ export const writeContractByFunctionName = async (functionName: string, ...args:
         throw formattedError(err)
     }
 }
+
+export const getContractEvents = async () => {
+    try {
+        return await client.getLogs({
+            address: contractAddress,
+            event: parseAbiItem(ProposalsRegistered),
+            fromBlock: BigInt(genesisBlock),
+            toBlock: 'latest'
+        })
+    } catch (err) {
+        throw formattedError(err)
+    }
+}
+
 
 const formattedError = (err: any): Error => {
     if (err instanceof BaseError) {